Split TRImage implementation
